XER5-916 : Dmcli support to initiate Xconf for XER5
Reason for change: Missing dmcli support to initiate Xconf for XER5
Test Procedure: Xconf should be triggered via Tr-181
Risks: Low
Priority: P1

(cherry picked from commit 5d400522a1e5d9c99d7bf1084ffa5365b9e0d470)

+++ source/TR-181/middle_layer_src/cosa_deviceinfo_dml.c
@@ -14978,6 +14978,11 @@ Xconf_SetParamBoolValue
                            v_secure_system ("kill -9 `pidof scer11bel_firmwareDwnld.sh `");
                        }
                            status = v_secure_system("/etc/scer11bel_firmwareDwnld.sh &");
+#elif defined(_XER5_PRODUCT_REQ_)
+			if(0 == v_secure_system("pidof xer5_firmwareDwnld.sh"))  {
+                           v_secure_system ("kill -9 `pidof xer5_firmwareDwnld.sh `");
+                       }
+                           status = v_secure_system("/etc/xer5_firmwareDwnld.sh &");
 #elif defined(_SR300_PRODUCT_REQ_)
                         if(0 == v_secure_system("pidof sr300_firmwareDwnld.sh"))  {
                            v_secure_system ("kill -9 `pidof sr300_firmwareDwnld.sh `");
